Move functions to separate file
